# nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "User related user stories API."
  version: "0.1.9"
  title: "GreenHero API"
...
paths:
...
  /Character:
    post:
      tags:
      - "Character"
      summary: "Create a new character"
      operationId: "createCharacter"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create a new character"
        required: true
        schema:
          $ref: "#/definitions/Character"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "Character"
    put:
      tags:
      - "Character"
      summary: "Modify character"
      description: "To modify a character "
      operationId: "modifyCharacter"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create a new character"
        required: true
        schema:
          $ref: "#/definitions/Character"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "character not found "
      x-swagger-router-controller: "Character"
  /Character/{user_id}:
    get:
      tags:
      - "Character"
      summary: "Get characters"
      description: "See the available character"
      operationId: "getCharacterByUserId"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "user_id"
        in: "path"
        description: "The user_id of the character."
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Character"
        "404":
          description: "No character found"
      x-swagger-router-controller: "Character"
  /Character/getCharacterByUserIdAndCharacterName:
    get:
      tags:
      - "Character"
      summary: "Get character"
      description: "See the available character"
      operationId: "getCharacterByUserIdAndCharacterName"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "The user_id of the character."
        required: true
        type: "string"
      - name: "characterName"
        in: "query"
        description: "The characterName of the character."
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Character"
        "404":
          description: "No character found"
      x-swagger-router-controller: "Character"
  /Character/deleteCharacterByUserIdAndCharacterName:
    delete:
      tags:
      - "Character"
      summary: "Delete character"
      description: "To delete a character"
      operationId: "deleteCharacterByUserIdAndCharacterName"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "The user_id of the character that needs to be deleted"
        required: true
        type: "string"
      - name: "characterName"
        in: "query"
        description: "The characterName of the character that needs to be deleted"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "character not found"
      x-swagger-router-controller: "Character"
  /Badge:
    get:
      tags:
      - "Badge"
      summary: "Get badges"
      description: "See the available badges"
      operationId: "getAllBadge"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "see the badges"
        required: true
        schema:
          $ref: "#/definitions/Badge"
      responses:
        "200":
          description: "OK"
        "404":
          description: "No badge found"
      x-swagger-router-controller: "Badge"
    post:
      tags:
      - "Badge"
      summary: "Create a new badge"
      operationId: "createBadge"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create a new badge"
        required: true
        schema:
          $ref: "#/definitions/Badge"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "Badge"
  /Badge/{badgeName}:
    put:
      tags:
      - "Badge"
      summary: "Modify badge"
      description: "To modify a badge "
      operationId: "modifyBadge"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "badgeName"
        in: "path"
        description: "The badgeName of the badge that needs to be modified"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated badge object"
        required: true
        schema:
          $ref: "#/definitions/Badge"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "badge not found "
      x-swagger-router-controller: "Badge"
    delete:
      tags:
      - "Badge"
      summary: "Delete badge"
      description: "To delete a badge"
      operationId: "deleteBadge"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "badgeName"
        in: "path"
        description: "The badgeName of the badge that needs to be deleted"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "badge not found"
      x-swagger-router-controller: "Badge"
...
  /Application:
    post:
      tags:
      - "Application"
      summary: "Create a new Application"
      operationId: "createApplication"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create a new application"
        required: true
        schema:
          $ref: "#/definitions/Application"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "Application"
    put:
      tags:
      - "Application"
      summary: "Modify Application"
      description: "To modify a Application"
      operationId: "modifyApplication"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Updated Application object"
        required: true
        schema:
          $ref: "#/definitions/Application"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "Application not found "
      x-swagger-router-controller: "Application"
  /Application/{team_id}:
    get:
      tags:
      - "Application"
      summary: "Get Applications by team ID"
      description: "See the available applications by team ID"
      operationId: "getAllApplicationsByTeamId"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "team_id"
        in: "path"
        description: "team_id  of application"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Application"
        "404":
          description: "No application found"
      x-swagger-router-controller: "Application"
  /Application/deleteApplicationByTeamIdAndUserId:
    delete:
      tags:
      - "Application"
      summary: "Delete Application"
      description: "To delete an application"
      operationId: "deleteApplicationByTeamIdAndUserId"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "team_id"
        in: "query"
        description: "team_id  of application"
        required: true
        type: "string"
      - name: "user_id"
        in: "query"
        description: "team_id  of application"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "No application found"
      x-swagger-router-controller: "Application"
...
  /AdminAnnounces:
    get:
      tags:
      - "AdminAnnounces"
      summary: "Get Admin Announces"
      description: "See the Admin Announces"
      operationId: "getAllAdminAnnounces"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AdminAnnounces"
        "404":
          description: "No temperature found"
      x-swagger-router-controller: "AdminAnnounces"
    post:
      tags:
      - "AdminAnnounces"
      operationId: "createAdminAnnounce"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create new admin announce"
        required: true
        schema:
          $ref: "#/definitions/AdminAnnounces"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "AdminAnnounces"
    put:
      tags:
      - "AdminAnnounces"
      summary: "Modify AdminAnnounces"
      description: "To modify a AdminAnnounces"
      operationId: "modifyAdminAnnounce"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "modify the AdminAnnounce"
        required: true
        schema:
          $ref: "#/definitions/AdminAnnounces"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "AdminAnnounce not found"
      x-swagger-router-controller: "AdminAnnounces"
  /AdminAnnounces/{_id}:
    delete:
      tags:
      - "AdminAnnounces"
      summary: "Delete adminAnnounce"
      operationId: "deleteAdminAnnounce"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "_id"
        in: "path"
        description: "The ID of the AdminAnnounces"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "AdminAnnounces not found"
      x-swagger-router-controller: "AdminAnnounces"
definitions:
...
  Character:
    type: "object"
    properties:
      _id:
        type: "string"
      user_id:
        type: "string"
      characterName:
        type: "string"
      characterDescription:
        type: "string"
      attack:
        type: "number"
      armor:
        type: "number"
      health:
        type: "number"
      healing_factor:
        type: "number"
      avatar:
        type: "string"
      skillTree:
        type: "string"
    xml:
      name: "Character"
    example:
      skillTree: "skillTree"
      armor: 6.02745618307040320615897144307382404804229736328125
      user_id: "user_id"
      attack: 0.80082819046101150206595775671303272247314453125
      characterDescription: "characterDescription"
      characterName: "characterName"
      health: 1.46581298050294517310021547018550336360931396484375
      _id: "_id"
      avatar: "avatar"
      healing_factor: 5.962133916683182377482808078639209270477294921875
  Badge:
    type: "object"
    required:
    - "badgeName"
    properties:
      _id:
        type: "string"
      badgeName:
        type: "string"
      badgeDescription:
        type: "string"
      badgePictureUrl:
        type: "string"
    xml:
      name: "Badge"
...
  Application:
    type: "object"
    properties:
      _id:
        type: "string"
      team_id:
        type: "string"
      user_id:
        type: "string"
      character:
        type: "string"
      accepted:
        type: "boolean"
      rejected:
        type: "boolean"
    xml:
      name: "Application"
    example:
      character: "character"
      user_id: "user_id"
      rejected: true
      accepted: true
      _id: "_id"
      team_id: "team_id"
...
  AdminAnnounces:
    type: "object"
    properties:
      _id:
        type: "string"
      icon:
        type: "string"
      title:
        type: "string"
      content:
        type: "string"
    example:
      icon: "icon"
      _id: "_id"
      title: "title"
      content: "content"
...